+++ packages/client-rpc/src/schema.ts
@@ -7,6 +7,19 @@ import {
   ProveResultSchema
 } from "./schema_elements.js";
 
+const CollectionIdSchema = v.string();
+
+const ProveArgsSchema = v.object({
+  request: PodspecProofRequestSchema,
+  collectionIds: v.optional(v.array(CollectionIdSchema))
+});
+
+export const PODDataSchema = v.object({
+  entries: PODEntriesSchema,
+  signature: v.string(),
+  signerPublicKey: v.string()
+});
+
 export type RPCFunction<
   I extends v.TupleSchema<
     v.BaseSchema<unknown, unknown, v.BaseIssue<unknown>>[],
@@ -38,15 +51,11 @@ type RPCSchema = Record<string, Record<string, RPCFunction>>;
 export const ParcnetRPCSchema = {
   gpc: {
     prove: {
-      input: v.tuple([PodspecProofRequestSchema] as [
-        proofRequest: typeof PodspecProofRequestSchema
-      ]),
+      input: v.tuple([ProveArgsSchema]),
       output: ProveResultSchema
     },
     canProve: {
-      input: v.tuple([PodspecProofRequestSchema] as [
-        proofRequest: typeof PodspecProofRequestSchema
-      ]),
+      input: v.tuple([ProveArgsSchema]),
       output: v.boolean()
     },
     verify: {
@@ -74,21 +83,31 @@ export const ParcnetRPCSchema = {
   },
   pod: {
     query: {
-      input: v.tuple([PODSchemaSchema] as [query: typeof PODSchemaSchema]),
-      output: v.array(v.string())
+      input: v.tuple([v.string(), PODSchemaSchema] as [
+        collectionId: ReturnType<typeof v.string>,
+        query: typeof PODSchemaSchema
+      ]),
+      output: v.array(PODDataSchema)
     },
     insert: {
-      input: v.tuple([v.string()] as [
-        serializedPOD: ReturnType<typeof v.string>
+      input: v.tuple([v.string(), PODDataSchema] as [
+        collectionId: ReturnType<typeof v.string>,
+        podData: typeof PODDataSchema
       ]),
       output: v.void()
     },
     delete: {
-      input: v.tuple([v.string()] as [signature: ReturnType<typeof v.string>]),
+      input: v.tuple([v.string(), v.string()] as [
+        collectionId: ReturnType<typeof v.string>,
+        signature: ReturnType<typeof v.string>
+      ]),
       output: v.void()
     },
     subscribe: {
-      input: v.tuple([PODSchemaSchema] as [query: typeof PODSchemaSchema]),
+      input: v.tuple([v.string(), PODSchemaSchema] as [
+        collectionId: ReturnType<typeof v.string>,
+        query: typeof PODSchemaSchema
+      ]),
       output: v.string()
     },
     unsubscribe: {
@@ -99,7 +118,7 @@ export const ParcnetRPCSchema = {
     },
     sign: {
       input: v.tuple([PODEntriesSchema] as [entries: typeof PODEntriesSchema]),
-      output: v.string()
+      output: PODDataSchema
     }
   }
 } as const satisfies RPCSchema;