+++ packages/client-helpers/CHANGELOG.md
@@ -1,5 +1,50 @@
 # @parcnet-js/client-helpers
 
+## 1.0.0
+
+### Major Changes
+
+- 2cc8bfc: 1.0.0 release
+
+  This release introduces two new features:
+
+  - Collections, which allow PODs to be stored in separate buckets with different access levels for different Zapps
+  - Permissions, which allow your Zapp to specify which features it requires access to
+
+  For documentation on these features, visit https://zappsdk.netlify.app or see the example code at `examples/test-app`.
+
+### Patch Changes
+
+- Updated dependencies [2cc8bfc]
+- Updated dependencies [a1539a5]
+  - @parcnet-js/client-rpc@1.0.0
+
+## 1.0.0-beta.1
+
+### Major Changes
+
+- 1.0.0 release
+
+  This release introduces two new features:
+
+  - Collections, which allow PODs to be stored in separate buckets with different access levels for different Zapps
+  - Permissions, which allow your Zapp to specify which features it requires access to
+
+  For documentation on these features, visit https://zappsdk.netlify.app or see the example code at `examples/test-app`.
+
+### Patch Changes
+
+- Updated dependencies
+  - @parcnet-js/client-rpc@1.0.0-beta.1
+
+## 0.0.9-alpha.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @parcnet-js/client-rpc@0.0.7
+  - @parcnet-js/client-rpc@0.0.7-alpha.0
+
 ## 0.0.8
 
 ### Patch Changes